requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
django-colorfield==0.12.0
django-tailwind[reload]
pytest
django-crispy-forms
crispy-tailwind